introduce use of config.json and components/interfaces.tsx to hold the interfaces
