+++ src/config.json
@@ -1,0 +1,7 @@
+{
+  "SERVER_URL": "https://example.com/api/",
+  "THEME_COLORS": {
+    "PRIMARY": "#007bff",
+    "SECONDARY": "#fc3"
+  }
+}
